Get person's age from database #57
